# src/main/swagger.json
{"swagger":"2.0","info":{"description":"Database query using Swaggerhub","version":"1.0","title":"Product API","termsOfService":"http://nous.com","license":{"name":"NousInfo License","url":"http://niweditar@nousinfo.com"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"product-controller","description":"Product Controller"}],"paths":{"/api/deleteById":{"delete":{"tags":["product-controller"],"summary":"deleteById","operationId":"deleteByIdUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"query","description":"id","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/api/getById":{"get":{"tags":["product-controller"],"summary":"getProductById","operationId":"getProductByIdUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"prodId","in":"query","description":"prodId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Product"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/getByName":{"get":{"tags":["product-controller"],"summary":"getProductByName","operationId":"getProductByNameUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"name","in":"query","description":"name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Product"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/getProduct":{"get":{"tags":["product-controller"],"summary":"findAllProducts","operationId":"findAllProductsUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Product"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/saveProduct":{"post":{"tags":["product-controller"],"summary":"save","operationId":"saveUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"product","description":"product","required":true,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Product"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/updateProd":{"put":{"tags":["product-controller"],"summary":"updateProd","operationId":"updateProdUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"query","description":"id","required":true,"type":"integer","format":"int32"},{"in":"body","name":"prod","description":"prod","required":true,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Product"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/updateProd1":{"put":{"tags":["product-controller"],"summary":"updateProd1","operationId":"updateProd1UsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"query","description":"id","required":true,"type":"integer","format":"int32"},{"in":"body","name":"prod","description":"prod","required":true,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Product"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Product":{"type":"object","properties":{"city":{"type":"string"},"emailId":{"type":"string"},"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}}}}